% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-functions.R
\name{apply_squared_exponential}
\alias{apply_squared_exponential}
\title{apply_squared_exponential}
\usage{
apply_squared_exponential(xi, xj, sig = 0.8, sig2 = 1)
}
\arguments{
\item{xi}{Corresponde ao valor numérico}

\item{xj}{Corresponde ao valor numperico}

\item{sig}{corresponde a um real}

\item{sig2}{corresponde a um número real}
}
\value{
A função retorna um numérico correspondendo a dois numéricos.
}
\description{
Métrica de similaridade utilizada para construir o grafo de similaridade.
A métrica utilizada é dada pela seguinte equação:
\deqn{S(i, j) = ||x_i - x_j||^2 / 2 x sigma1 x sigma 2}
}
\examples{
apply_squared_exponential(7, 9)
}
