% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-functions.R
\name{build_similarity_graph}
\alias{build_similarity_graph}
\title{build_similarity_graph}
\usage{
build_similarity_graph(A, sig1 = 0.8, sig2 = 1)
}
\arguments{
\item{A}{matriz numérica quadrada}

\item{sig1}{corresponde a um real utilizado na função de similaridae}

\item{sig2}{corresponde a um real utilizado na função de similaridae}
}
\value{
A função retorna uma matriz correspondendo a grafo de similaridade.
}
\description{
Essa função realiza a transformação da matriz em um grafo de similaridade.
Usando a função apply_squared_exponential para obter a similaridade.
}
\examples{
A <- matrix(sample(100), nrow = 10)
print(A)
res <- build_similarity_graph(A)
print(res)

}
