% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-functions.R
\name{create_graph_laplacian}
\alias{create_graph_laplacian}
\title{create_graph_laplacian}
\usage{
create_graph_laplacian(similarity_matrix, type = 1)
}
\arguments{
\item{similarity_matrix}{corresponde a matriz de similaridade calculada.}

\item{type}{Corresponde a um inteiro indicando qual tipo de normalização deseja
executar.
\itemize{
\item 1: Não normalizado.
\item 2: Normalizado segundo Shi e Malik (2000).
\item 3: Normalizado segundo Ng, Jordan, e Weiss (2002).
}}
}
\value{
A função retorna uma matriz correspondendo a grafo de laplaciano.
}
\description{
Essa função realiza o cálculo do grafo Laplaciano, o qual pode ser não normalizado, ou
normalizado segundo Shi e Malik (2000) ou Ng, Jordan, e Weiss (2002).
}
\examples{
set.seed(2018)
x <- sample(100, 10)
A <- as.matrix(dist(x))
S <- build_similarity_graph(A)
L <- create_graph_laplacian(S)

}
