% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateData}
\alias{generateData}
\title{generateData}
\usage{
generateData(type, islandsIn = F, states = vector(), conFactor = 1)
}
\arguments{
\item{type}{three options "dataTerr", "dataFW", and "dataTerrFW"}

\item{states}{a vector of states names that have to be included}

\item{conFactor}{contiguity constraint factor}

\item{islandIn}{if T the islands will be included}

\item{islands}{vector of islands?}
}
\value{
a list with three elements: data, conMatrix, and latLong
}
\description{
Generate the data for clustering
}
\examples{
\dontrun{
dataTerr     <- read.csv("data/terrData.csv", header = T)
dataFW       <- read.csv("data/freshData.csv", header = T)
i <- which(colnames(dataFW) == "hu12_states")
dataTerrFW   <- merge(dataTerr, dataFW[-i], by.x = "zoneid", by.y = "zoneid")
islands      <- read.csv("data/islandIdx.csv", header = T)

input <- generateData(type = dataTerrFW, islands, islandsIn = F,
         states = c("MO"),conFactor=1)
}
}

