% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hspeCluster}
\alias{hspeCluster}
\title{hspecCluster}
\usage{
hspeCluster(data, conMatrix, cluster.number, iter.max = 400,
  repetition = 400)
}
\arguments{
\item{data}{A numeric data frame or matrix where columns represent variables and rows represent locations}

\item{conMatrix}{Contiguity matrix.}

\item{cluster.number}{The number of clusters.}

\item{iter.max}{The maximum number of iterations allowed for kmeans step.}

\item{repetition}{How  many  random  sets  should  be  chosen for  as  the  initial centers in kmeans step.}
}
\description{
hspecCluster
}
\examples{
\dontrun{
load("HU12SpectralClustering.RData")
input <- generateData(type = dataTerrFW, islandsIn = F, states = c("MI"), conFactor=1)
results <- hspeCluster(data= input$data, conMatrix=input$conMatrix, cluster.number=8)
mapping(lat = input$latLong[,1],long=input$latLong[,2],clusters= results[,1]$clusters)
}
}
