% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess.R
\name{mapping}
\alias{mapping}
\title{mapping}
\usage{
mapping(long, lat, clusters)
}
\arguments{
\item{long}{A numeric vector of  longitude location of the points.}

\item{lat}{A numeric vector of latitude location of the points.}

\item{clusters}{The  vector of integers indicating the cluster to which each point is allocated.}
}
\description{
Using map database draw cluster on the US Map
}
