% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nb_collapse}
\alias{nb_collapse}
\title{nb_collapse}
\usage{
nb_collapse(nb)
}
\arguments{
\item{nb}{neighbor list}
}
\value{
neighbor data.frame
}
\description{
convert a neighbor list to a data.frame
}
\examples{
\dontrun{
nb_c <- nb_collapse(nb)
}
}

