% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{neighborMatrix}
\alias{neighborMatrix}
\title{neighborMatrix}
\usage{
neighborMatrix(NB, conFactor = 1)
}
\arguments{
\item{NB}{data.frame neighbor list}

\item{conFactor}{integer contiguity constraint factor}
}
\description{
Compute constraint Matrix
}

