% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{outlierDetector}
\alias{outlierDetector}
\title{outlierDetector}
\usage{
outlierDetector(data, outlier.Threshold = 0.2)
}
\arguments{
\item{data}{data.frame}

\item{outlier.Threshold}{numeric}
}
\description{
outlierDetector
}
