% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralClustering.R
\name{similarity}
\alias{similarity}
\title{Similarity}
\usage{
similarity(data, neighbors)
}
\arguments{
\item{data}{n by p numeric matrix or data frame.}

\item{neighbors}{n by n numeric matrix which specifies contiguity matrix.}
}
\value{
similarity: The similarity matrix
}
\description{
Compute similarity matrix then multiply it to contiguity matrix
}
