% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-functions.R
\docType{data}
\name{spec_data}
\alias{spec_data}
\title{speclusteR: conjunto para teste}
\format{An object of class \code{list} of length 2.}
\usage{
spec_data
}
\description{
Lista representando dois conjuntos de pontos utilizados para teste
list(espirais, happy)
}
\keyword{datasets}
