% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-functions.R
\docType{package}
\name{speclusteR}
\alias{speclusteR}
\alias{speclusteR-package}
\title{speclusteR: Spectral Clustering Algorithms in R}
\description{
Trabalho para disciplina Álgebra Linear para Data Science Machine Learning - 2018.2,
implementação dos três principais Spectral Clustering Algorithms.
}
\references{
\itemize{
\item https://bit.ly/2B6qFnn
\item https://bit.ly/2zRckeK
\item https://bit.ly/2G5UmL0
}
}
\author{
Jessica Cardoso
}
