% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stepOne}
\alias{stepOne}
\title{stepOne}
\usage{
stepOne(data, conMatrix, ncol)
}
\arguments{
\item{data}{data.frame}

\item{conMatrix}{contiguity matrix}

\item{ncol}{number of output columns}
}
\description{
stepOne
}
