% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess.R
\name{sumSquares}
\alias{sumSquares}
\title{sumSquares}
\usage{
sumSquares(data, clusters)
}
\arguments{
\item{data}{after principal component data}

\item{clusters}{the vector of integers indicating cluster membership}
}
\description{
sumSquares
}
