% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{speCluster}
\alias{speCluster}
\title{speCluster}
\usage{
speCluster(data, conMatrix, cluster.number, iter.max = 400,
  repetition = 400)
}
\arguments{
\item{data}{A numeric data frame or matrix.}

\item{conMatrix}{Contiguity matrix.}

\item{cluster.number}{The number of clusters.}

\item{iter.max}{The maximum number of iterations allowed for kmeans step.}

\item{repetition}{How  many  random  sets  should  be  chosen for  as  the  initial centers in kmeans step.}
}
\value{
A list contains two parts:
        clusters: A vector of integers(from 1:cluster.number)
        indicating the cluster to which each point is allocated.
       SS: A list with two values SSW for Sum Squered Within and
                 SSB for SumSquered Between
}
\description{
Perform Spectral Clustering on a data matrix
}
\examples{
\dontrun{
speCluster()
}
}

